Added functionality for having different skills for different indies. Kinda bugs out when cancelling though. Changed MobStats to have the same byte order as CharacterTemporaryStats. Added a ton of InHeaders. Added handlers for PsychicArea/Lock Creation/Release. Fixed user damage by mob. Added better functionality for summons. Just Field.spawnSummon() now. Fixed summons not despawning (either by force or by timeout).
Classes:

Kinesis:
Fixed PP system.
Fixed Psychic Assault/Blast.
Fixed Mind Break.
Fixed Kinetic Combo creation. Does not yet do damage, unknown cause atm.
Fixed Psychic Armor/Bulwark.
Fixed Pure Power.
Fixed Psychic Reinforcement.
Fixed President's Orders.
Fixed Telepath Tactics.
Fixed Psychic Charger.
Kinetic Jaunt 38s.
Gotta figure out mental shield. Half-works atm.

Ice/Lightning:
Fixed Magic Guard.
Fixed IL's skills that were equal to FP's (medidation, booster, etc...)
Fixed IL's stun + freeze + summon skills.
Fixed !lookup command not getting the entire query.

+++ src/client/character/Char.java
@@ -1290,7 +1290,7 @@ public class Char {
         return skill;
     }
     
-    public void setStat(Stat charStat, short amount) {
+    public void setStat(Stat charStat, int amount) {
         switch(charStat) {
             case str:
                 getAvatarData().getCharacterStat().setStr(amount);
@@ -1325,7 +1325,7 @@ public class Char {
         }
     }
     
-    public short getStat(Stat charStat) {
+    public int getStat(Stat charStat) {
         switch(charStat) {
             case str:
                 return getAvatarData().getCharacterStat().getStr();
@@ -1349,8 +1349,8 @@ public class Char {
         return -1;
     }
     
-    public void addStat(Stat charStat, short amount) {
-        setStat(charStat, (short) (getStat(charStat) + amount));
+    public void addStat(Stat charStat, int amount) {
+        setStat(charStat, getStat(charStat) + amount);
     }
 
     public Position getOldPosition() {
